Heavy changes ... mbag uses const char * as key
WTP is still not 100% working...

FossilOrigin-Name: 9e80cb8f1b62f854b54393e3af75b789125870b1397ccccad6d3442e3395181c

+++ src/capwap/conn_process_packet.c
@@ -262,7 +262,7 @@ static int process_elements(struct conn *conn, uint8_t * rawmsg, int len,struct 
 		if (af->mand && afrc) {
 			/* add found mandatory message element 
 			   to mand list */
-			intavltree_add(conn->mand, af->item_id);
+			intavltree_add(conn->mand, (int)af->item_id);
 		}
 
 	}
